requirements.txt
aiohttp==3.7.4.post0
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.5
beautifulsoup4==4.9.3
bottle==0.12.19
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
discord.py==1.6.0
Flask==1.1.2
future==0.18.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
line-bot-sdk==1.19.0
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
oauthlib==3.1.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
PyNaCl==1.4.0
PySocks==1.7.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
yarl==1.6.3
